+++ CHANGELOG.md
@@ -32,7 +32,7 @@ Mark  items as `Added`, `Changed`, `Fixed`, `Removed`, `Untested Features`, `Upc
 
 **forms-flow-idm**
 
-* Added manage-users group to formsflow-ai-realm.
+* Added `manage-users` group to assigned client roles in realm-management.Check out the details at Service Accounts Tab from [here](./forms-flow-idm/keycloak/README.md#create-a-forms-flow-bpm-client).
 
 **forms-flow-bpm**